pyproject.toml
[tool.poetry]
name = "debt_bot"
version = "0.1.0"
description = "Debt bot for two people on Telegram"
authors = ["sWallyx <mikelsmartinez@aol.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
python-telegram-bot = "^13.8.1"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
